replace method pablish learn queue
in Learn manager replace method next to publisher combine
clean code
