+++ FlashSpeak/FlashSpeak/Flow/Study/Learn/Presenter/LearnPresenter.swift
@@ -6,10 +6,12 @@
 //
 
 import UIKit
+import Combine
 
 protocol LearnViewInput {
     var question: Question { get set }
     var answer: Answer { get set }
+    /// Required to bind a delegate to a cell
     var answerTextFieldDelegate: UITextFieldDelegate { get }
     
     /// Initial configure answer view by type
@@ -35,7 +37,6 @@ protocol LearnViewOutput {
     /// Answer received. Checking the answer by type and highlighting the answer view.
     func didAnsewred(answer: Answer)
     /// Change question and answer. Get exercise from manager and set to view
-    func nextQuestion()
 }
 
 class LearnPresenter {
@@ -43,26 +44,32 @@ class LearnPresenter {
     // MARK: - Properties
     
     var list: List
-    var session: LearnSession
-    
     var router: LearnEvent
     weak var viewController: (UIViewController & LearnViewInput)?
     
     // MARK: - Private properties
+    
+    private var manager: LearnManager
+//    private var store = Set<AnyCancellable>()
+//    private let subject = PassthroughSubject<Exercise, Never>()
     
     // MARK: - Constracions
     
     init(router: LearnEvent, list: List, settings: LearnSettings) {
         self.list = list
         self.router = router
-        self.session = LearnSession(words: list.words, settings: settings)
-        session.delegate = self
+        self.manager = LearnManager(words: list.words, settings: settings)
+        manager.delegate = self
     }
     
     // MARK: - Private functions
     
     private func configureAnswerView() {
-        viewController?.configureAnswerView(settings: session.settings.answer)
+        viewController?.configureAnswerView(settings: manager.settings.answer)
+    }
+    
+    private func start() {
+        manager.start()
     }
 }
 
@@ -70,41 +77,38 @@ extension LearnPresenter: LearnViewOutput {
     
     func getConfigure() {
         configureAnswerView()
-        nextQuestion()
-    }
-    
-    func nextQuestion() {
-        if let exercise = session.next() {
-            viewController?.configure(exercise: exercise, settings: session.settings.answer)
-        }
+        start()
     }
     
     func didAnsewred(answer: Answer) {
-        switch session.settings.answer {
+        switch manager.settings.answer {
         case .test:
-            let isRight = session.isRight(userAnswer: answer)
-            let rightIndex = session.rightIndexTest()
-            let userIndex = session.indexTest(userAnswer: answer)
+            let isRight = manager.isRight(userAnswer: answer)
+            let rightIndex = manager.rightIndexTest()
+            let userIndex = manager.indexTest(userAnswer: answer)
             if !isRight {
                 viewController?.highlightAnswer(isRight: false, index: userIndex, settings: .test)
             }
             viewController?.highlightAnswer(isRight: true, index: rightIndex, settings: .test)
         case .keyboard:
-            let isRight = session.isRight(userAnswer: answer)
+            let isRight = manager.isRight(userAnswer: answer)
             if isRight {
                 viewController?.highlightAnswer(isRight: true, index: .zero, settings: .keyboard)
             } else {
                 viewController?.highlightAnswer(isRight: false, index: .zero, settings: .keyboard)
             }
         }
-        session.answered()
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.6) {
-            self.nextQuestion()
+            self.manager.answered()
         }
     }
 }
 
-extension LearnPresenter: LearnSessionDelegate {
+extension LearnPresenter: LearnManagerDelegate {
+    
+    func receive(exercise: Exercise, answer: LearnSettings.Answer) {
+        viewController?.configure(exercise: exercise, settings: answer)
+    }
     
     func complete(learn: Learn) {
         print(#function, learn)
